refactor(hooks): removes standalone `hooks` module
We didn't have many "external" hooks (only some "internal" ones), and still, it would be hard to "set" a convention for users (e.g. to use signals or hooks). Also, the full-stack nature of fresh could cause confusion as to when to use hooks for users. Users can obviously still make use of regular preact hooks (or signals if they prefer) and external hook libraries.
