+++ lib/components/toaster.tsx
@@ -6,7 +6,7 @@ import {
   ToastTitle,
   ToastViewport,
 } from "./toast.tsx";
-import { useToast } from "../hooks/use-toast.ts";
+import { useToast } from "./use-toast.ts";
 
 export function Toaster() {
   const { toasts } = useToast();
