+++ lib/components/use-toast.ts
@@ -1,6 +1,6 @@
 // Inspired by react-hot-toast library
 
-import type { ToastActionElement, ToastProps } from "../components/toast.tsx";
+import type { ToastActionElement, ToastProps } from "./toast.tsx";
 import type { VNode } from "../deps/preact.ts";
 import { useEffect, useState } from "../deps/preact/compat.ts";
