+++ lib/components/utils.ts
@@ -1,3 +1,4 @@
+import { deepMerge } from "../deps/std/collections/deep_merge.ts";
 // unocss might have twMerge alternative soon
 // see https://github.com/unocss/unocss/issues/2748
 import { twMerge } from "../deps/tailwind-merge.ts";
@@ -9,3 +10,28 @@ export const cn = (...inputs: ClassValue[]) => {
 };
 
 export const twx = createTwc({ compose: cn });
+
+/* Internal utility used by netzo for components */
+export const useUI = <T = Record<string, unknown>>(
+  ui: Partial<T>,
+  config: Partial<T> = {},
+): T => {
+  const keys = [...new Set([...Object.keys(ui), ...Object.keys(config)])];
+  return keys.reduce((acc, key) => {
+    const {
+      class: configClass,
+      className: configClassName,
+      ...configRest
+    } = (config as any)?.[key] ?? {};
+    const {
+      class: uiClass,
+      className: uiClassName,
+      ...uiRest
+    } = (ui as any)?.[key] ?? {};
+    const className = cn(configClass, configClassName, uiClass, uiClassName);
+    return {
+      ...acc,
+      [key]: { ...deepMerge(configRest, uiRest), className },
+    };
+  }, {} as T) as T;
+};
