+++ lib/deps/usehooks-ts.ts
@@ -1,0 +1,1 @@
+export * from "https://esm.sh/usehooks-ts@2.14.0?external=react,react-dom&target=es2022";
