--- lib/hooks/server/use-netzo-state.ts
@@ -1,11 +1,0 @@
-// import { IS_BROWSER } from "../../deps/$fresh/runtime.ts";
-import type { NetzoState } from "../../mod.ts";
-
-export function useNetzoState(ctx: FreshContext<NetzoState>) {
-  const { auth } = ctx.state.config;
-  const { sessionId, sessionUser } = ctx.state.auth ?? {};
-
-  const mustAuth = !!auth && !sessionId;
-
-  return { sessionId, sessionUser, mustAuth };
-}
